+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Changelog for NeoFS Node
 - Token and object authentication errors are more detailed in status responses now (#3216, #3264)
 - IR uses NNS names for alphabet contracts now instead of Glagolitsa (#3268)
 - neofs-adm can handle any wallet names in its configuration (#3268)
+- neofs-adm uses alphabet0/1/2 names for wallet files by default now (#3268)
 
 ### Removed
 - SN `apiclient.allow_external` config (#3235)